+++ lib/stat_power.rb
@@ -7,6 +7,7 @@ require_relative "stat_power/unequal_power_result"
 require_relative "stat_power/anova_result"
 require_relative "stat_power/f2_result"
 require_relative "stat_power/chi_square_result"
+require_relative "stat_power/power_curve_point"
 require_relative "stat_power/errors"
 require_relative "stat_power/special_functions/beta"
 require_relative "stat_power/special_functions/gamma"
@@ -30,6 +31,7 @@ require_relative "stat_power/correlation"
 require_relative "stat_power/anova"
 require_relative "stat_power/f2"
 require_relative "stat_power/chi_square"
+require_relative "stat_power/power_curve"
 
 module StatPower
 end